Android - Ok

+++ shared/src/commonMain/kotlin/com/works/data/dto/UserLoginResponseDto.kt
@@ -8,19 +8,19 @@ import kotlinx.serialization.Serializable
 )
 
 @Serializable data class Meta(
-    val status: Long,
+    val status: Int,
     val message: String,
 )
 
 @Serializable data class Data(
     val access_token: String,
     val token_type: String,
-    val expires_in: Long,
+    val expires_in: Int,
     val user: User,
 )
 
 @Serializable data class User(
-    val id: Long,
+    val id: Int,
     val name: String,
     val email: String,
     val role: String,
